% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_autostart}
\alias{dropbox_autostart}
\title{Enable or Disable the autostartup feature}
\usage{
dropbox_autostart(autostart = T)
}
\arguments{
\item{autostart}{A \code{boolean} indicating \code{TRUE} (enabled)
or \code{FALSE} (disabled) autostart of the dropbox daemon.}
}
\description{
Tries to enable autostartup when a user is detected as logged in.
}
\details{
This is feature is required for the Dropbox CLI package to work as R
will otherwise close connections quickly.
}
\examples{
\dontrun{
# Enable Autostart
dropbox_autostart(T)

# Disable Autostart
dropbox_autostart(F)
}
}
