% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_command}
\alias{dropbox_command}
\title{Issue Dropbox Commands to System}
\usage{
dropbox_command(cmd, dbpy_loc = getOption("dropcli.cmdloc", "~/"),
  intern = TRUE)
}
\arguments{
\item{cmd}{A \code{string} containing the command and option.}

\item{dbpy_loc}{A \code{string} indicating the location of
the \code{"dropbox.py"} CLI interface.}

\item{intern}{A \code{bool} that represents whether output should
be returned as a \code{character vector} or not.}
}
\description{
Enables the use of the system terminal's
\code{"dropbox.py"} CLI interface.
}
\examples{
\dontrun{
# Start process
dropbox_command("start")
}
}
