% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_help}
\alias{dropbox_help}
\title{Dropbox Help Command}
\usage{
dropbox_help(topic = NULL)
}
\arguments{
\item{topic}{A \code{string} containing a valid command or \code{NULL}}
}
\description{
Retrieves the present daemon status
}
\details{
Valid \code{topic}s are:
\tabular{rr}{
status     \tab   get current status of the dropboxd                      \cr
throttle   \tab   set bandwidth limits for Dropbox                        \cr
help       \tab   provide help                                            \cr
puburl     \tab   get public url of a file in your dropbox's public folder\cr
stop       \tab   stop dropboxd                                           \cr
running    \tab   return whether dropbox is running                       \cr
start      \tab   start dropboxd                                          \cr
filestatus \tab   get current sync status of one or more files            \cr
ls         \tab   list directory contents with current sync status        \cr
autostart  \tab   automatically start dropbox at login                    \cr
exclude    \tab   ignores/excludes a directory from syncing               \cr
lansync    \tab   enables or disables LAN sync                            \cr
sharelink  \tab   get a shared link for a file in your dropbox            \cr
proxy      \tab   set proxy settings for Dropbox                          \cr
}

If \code{NULL}  is supplied then the above index is printed.
}
\examples{
\dontrun{
# Get standard help table
dropbox_help()

# Request help with puburl
dropbox_help("puburl")
}
}
