% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_running}
\alias{dropbox_running}
\title{Is Dropbox Running?}
\usage{
dropbox_running()
}
\value{
A \code{bool} containing either \code{TRUE} (running) or
\code{FALSE} (not running)
}
\description{
Determines whether or not the dropbox daemon is running.
}
\examples{
\dontrun{
# Check to see if dropbox is running
dropbox_running()
}
}
