% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_sharelink}
\alias{dropbox_sharelink}
\title{Obtain the File URL in a Public Folder}
\usage{
dropbox_sharelink(fp)
}
\arguments{
\item{fp}{A \code{string} containing the file path that is within the \code{"Public"} folder.}
}
\description{
Returns the Public File URL in Dropbox's public folder
}
\details{
The file must be located \strong{outside} of the \code{Public} directory
}
\examples{
\dontrun{
# Return the file
dropbox_sharelink("path/to/file.R")
}
}

