% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{dropbox_status}
\alias{dropbox_status}
\title{Dropbox Daemon Status}
\usage{
dropbox_status()
}
\description{
Retrieves the present daemon status
}
\examples{
\dontrun{
# Daemon status
dropbox_status()
}
}
