% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropcli-package.r
\docType{package}
\name{dropcli-package}
\alias{dropcli}
\alias{dropcli-package}
\title{dropcli: Interface with Linux Dropbox CLI}
\description{
Provides the ability to use R to interface with the
Linux Dropbox Command Line Interface.
}
\details{
The package is in its initial stage and is not ready for development
}
\author{
\strong{Maintainer}: James Balauta \email{balamut2@illinois.edu}

}
